[mlir] move MatchOpInterface under Transform/Interfaces
This is similar to the TransformOpInterface move.
