+++ mlir/lib/Dialect/Transform/IR/CMakeLists.txt
@@ -1,14 +1,9 @@
 add_mlir_dialect_library(MLIRTransformDialect
-  MatchInterfaces.cpp
   TransformAttrs.cpp
   TransformDialect.cpp
   TransformOps.cpp
   TransformTypes.cpp
   Utils.cpp
-
-  DEPENDS
-  MLIRMatchInterfacesIncGen
-  MLIRTransformDialectIncGen
 
   LINK_LIBS PUBLIC
   MLIRCastInterfaces
